updateNew

+++ laravelsi4b/example-app/app/Http/Controllers/MahasiswaController.php
@@ -12,9 +12,15 @@ class MahasiswaController extends Controller
     /**
      * Display a listing of the resource.
      */
+    
     public function index()
     {
-        $mahasiswa = Mahasiswa::all();
+        if(auth()->user()->role == 'D'){
+            $mahasiswa = Mahasiswa::where('user_id',auth()->user()->id)->get();
+
+        } else {
+            $mahasiswa = Mahasiswa::all();
+        }
         return view('mahasiswa.index')
                 ->with('mahasiswa',$mahasiswa);
     }
@@ -33,6 +39,9 @@ class MahasiswaController extends Controller
      */
     public function store(Request $request)
     {
+        if($request->user()->cannot('create',Mahasiswa::class)){
+            abort(403);
+        }
          // dd($request);
         // validasi form
         $val = $request->validate([
@@ -57,7 +66,7 @@ class MahasiswaController extends Controller
         // redirect ke halaman list fakultas
         return redirect()->route('mahasiswa.index')->with('Success', $val['nama'] . ' berhasil disimpan');
     }
-
+    
     /**
      * Display the specified resource.
      */
@@ -83,6 +92,9 @@ class MahasiswaController extends Controller
      */
     public function update(Request $request, Mahasiswa $mahasiswa)
     {
+        if(auth()->user()->cannot('update',$mahasiswa)){
+            abort(403);
+        }
         // dd($mahasiswa);
         if($request->url_foto){ // jika da file foto yang dilampirkan
             $val = $request->validate([
@@ -123,6 +135,9 @@ class MahasiswaController extends Controller
      */
     public function destroy(Mahasiswa $mahasiswa)
     {
+        if(auth()->user()->cannot('delete',$mahasiswa)){
+            abort(403);
+        }
         // dd($mahasiswa);
         File::delete('foto/'. $mahasiswa['url_foto']);
         $mahasiswa->delete(); // hapus data mahasiswa
